refactor: Cambiar tipo de ID en repositorios y añadir métodos de paginación para Usuario

+++ src/main/java/com/cibertec/demo/repository/CargaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CargaRepository extends JpaRepository<Carga, Integer> {
+public interface CargaRepository extends JpaRepository<Carga, Long> {
 }
