+++ src/main/java/com/cibertec/demo/repository/DetalleCitaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DetalleCitaRepository extends JpaRepository<DetalleCita, Integer> {
+public interface DetalleCitaRepository extends JpaRepository<DetalleCita, Long> {
 }
