+++ src/main/java/com/cibertec/demo/repository/OperadorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OperadorRepository extends JpaRepository<Operador, Integer> {
+public interface OperadorRepository extends JpaRepository<Operador, Long> {
 }
